#246 parsePositionMeta on the wrong file
https://github.com/MoreMoney-Finance/frontend/issues/246
